pyproject.toml
[project]
name = "ai-dac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diagrams",
    "boto3",
    "PyYAML",
    "GitPython",
]
